feat(backend): removed video-storage and splitted logic
